+++ pkg/ebpf/co-re/struct_flavors.h
