+++ pkg/ebpf/events_derived.go
@@ -1,4 +1,4 @@
-package tracee
+package ebpf
 
 import (
 	"fmt"
