+++ pkg/ebpf/events_pipeline.go
@@ -1,4 +1,4 @@
-package tracee
+package ebpf
 
 import (
 	"bytes"
