+++ pkg/ebpf/events_processor.go
@@ -1,4 +1,4 @@
-package tracee
+package ebpf
 
 import (
 	"fmt"
