+++ pkg/ebpf/tracee.go
@@ -1,4 +1,4 @@
-package tracee
+package ebpf
 
 import (
 	gocontext "context"
