+++ pkg/ebpf/user_mode_events_test.go
@@ -1,4 +1,4 @@
-package tracee
+package ebpf
 
 import (
 	"testing"
